requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
boto3==1.34.50
botocore==1.34.50
certifi==2024.2.2
charset-normalizer==3.3.2
dataclasses-json==0.6.4
distro==1.9.0
frozenlist==1.4.1
greenlet==3.0.3
grpcio==1.62.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.27
langsmith==0.1.9
marshmallow==3.21.0
momento==1.19.1
momento-wire-types==0.106.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.12.0
orjson==3.9.15
packaging==23.2
pinecone-client==3.1.0
protobuf==4.25.3
pydantic==2.6.2
pydantic_core==2.16.3
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
slack-bolt==1.18.1
slack_sdk==3.27.0
sniffio==1.3.1
SQLAlchemy==2.0.27
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.0.7
yarl==1.9.4
